sf: fix SurfaceView Dim layer show black issue
Gallery play video, press back key to end the playing, the sceen
flash black. SurfaceView Dim is the child of SurfaceView, but the
alphia do not change with the parent layer at the exiting. Thus is
the SurfaceView Dim layer do change the alpha along with the
transition animation. getAlpha can fix this issue, getAlpha with
check the parent alpha.

Test: make, take a photo, record a video and play video

Change-Id: I8b91682e33c2f6c1f52a3da25e051dcb58e99dc1
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: spezi77 <spezi77@gmx.com>